refactor generate env tool
