+++ helper-scripts/src/stackManagementTool/generateEnv/generateEnvTool.ts
@@ -1,12 +1,18 @@
-import { generateEnvForMainStacks } from "./generateEnvForMainStacks/generateEnvForMainStacks.js";
-import { generateEnvForCustomStacks } from "./customStacks/generateEnvForCustomStacks.js";
-import { askCustomStackOrMainDev } from "./prompts.js";
+import {
+  askCustomStackOrMainDev,
+  whichStacksToGenerateEnvFor,
+} from "./prompts.js";
+import { echo } from "zx";
+import { generateEnvs } from "./getStackName/generateEnvs/generateEnvs.js";
+import { getStackName } from "./getStackName/getStackName.js";
 
 export const generateEnvTool = async (): Promise<void> => {
   const { customStackOrMain } = await askCustomStackOrMainDev();
-  if (customStackOrMain === "Main stack(s) in dev") {
-    await generateEnvForMainStacks();
-  } else {
-    await generateEnvForCustomStacks();
-  }
+  echo("");
+
+  const baseStackName = await getStackName(customStackOrMain);
+  echo("");
+
+  const { choice } = await whichStacksToGenerateEnvFor(baseStackName);
+  await generateEnvs(baseStackName, choice);
 };
